refactor: lift TypedDispatcher to core/utils + extract custom-command config (Phase 10)
Two extracts from Phase 10's audit:

1. createDispatcher<U, Args>() lifted from messageHandlers.ts into
   src/core/utils/dispatcher.ts as a generic utility. Same pattern
   messageHandlers proved out (DispatcherMap with exhaustive
   { [K in U['type']]: Handler<K> } indexing), but reusable from
   workflow-editor, spec-editor, and any future webview surface.
   messageHandlers.ts now does:
     const dispatch = createDispatcher(buildHandlerMap());
     await dispatch(message, specDirectory, deps);
   ...with no local Handler<K> type, no manual cast plumbing.

2. The custom-command resolution helpers (NormalizedCustomCommand type,
   loadCustomCommands, normalizeCustomCommand, buildCustomCommand) moved
   from specCommands.ts into a dedicated customCommandConfig.ts. They
   were ~80 LOC of pure normalisation logic that didn't depend on the
   command-registration plumbing they were nested inside — natural cut.
   specCommands.ts is now 834 → 756 LOC.

The bigger specCommands extract (audit's "specCommandRegistry" idea) is
deferred. The current closure-with-shared-deps shape works; the file is
large but coherent and the boundaries between command groups aren't
obvious enough to justify an aggressive split.

All 729 tests pass. Net production-code delta: −13 messageHandlers,
−78 specCommands, +85 new utility + new helper file.
